Migrated investigation selectors into a new file to consolidate selectors because addition of new selectors in instruments selector file is causing circular dependencies. Other selector files will be consolidated in subsequent commits.
