+++ apps/frontend/src/pages/instruments/detail.tsx
@@ -15,7 +15,7 @@ import InvestigationStatus from "src/components/InvestigationStatus/Investigatio
 import { PDS4_INFO_MODEL } from "src/types/pds4-info-model";
 import { Stats, StatsList } from "src/components/StatsList/StatsList";
 import { selectLatestInstrumentHostVersion } from "src/state/selectors/instrumentHost";
-import { selectLatestInvestigationVersion } from "src/state/selectors/investigations";
+import { selectLatestInvestigationVersion } from "src/state/selectors";
 import FeaturedDataBundleLinkListItem from "src/components/FeaturedListItems/FeaturedDataBundleLinkListItem";
 import { ArrowOutward } from "@mui/icons-material";
